--- components/About.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-export const About = () => {
-  return (
-      <div className='container'>
-        <h2>About Me</h2>
-          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit ut aliquam, purus sit amet luctus 
-          venenatis, lectus magna fringilla urna, porttitor rhoncus dolor purus non enim praesent  elementum facilisis leo, vel  fringilla est ullamcorper  eget nulla facilisi etiam dignissim  diam quis enim lobortis  scelerisque fermentum dui 
-          faucibus in ornare quam viverra </p>
-      </div>
-  )
-};
